Added console compatibility to commands
it's a bit frustrating to have to be in game to use commands that should be available from the console. This commit adds the ability for the console to execute various commands.

Commands that console can now use:
- kick
- ban
- unban
- slap
- setgametype
- setmapname

I've also added a utility function "getName" which, returns either the player's name of the passed source, "Console" or "Unknown". The function should only return "Unknown" if the passed number is not 0 and the name isn't returned by the "GetPlayerName" native.

I've also modified the "DoesPlayerHavePermission" function to return true if the "player" is console.