refactor search + add notification for no results
*This update will prevent some edge cases*
if someone searched by ID and there happened to be a player w/ that number in their name, it wouldn't work before.
*This update adds convenience*
If only a single result is found, will now jump straight to that player's menu!
*Also added notification*
If  no results are found, you get a notification so users don't think it's just broken.