add Keywords
Essentially, if the server owner adds ea_addKeyword <KEYWORD> <Thing to replace that with> to their cfg, it will replace that with a pre-typed message, but ONLY if that is the entire reason for a kick/ban/warn.
I.E.
Cfg -> ea_addKeyword VDM Vehicle Deathmatch is not allowed, please read our rules!
Admin -> *warns player* reason: vdm
User -> Sees the "Vehicle Deathmatch is not allowed, please read our rules!" message