patch(exploit/sanitization) injecting html img into calladmin
Added sanitization check to reason on calladmin and ban player attempting to place a url in the reason. Players can place an image on their own server and then input the url to grab player IPs or display FULL SCREEN images on all Staffs Screens. Added check and ban the injector